+++ core/chains/evm/gas/gas_test.go
@@ -25,7 +25,7 @@ func Test_BumpLegacyGasPriceOnly(t *testing.T) {
 		bumpWei                *big.Int
 		maxGasPriceWei         *big.Int
 		expectedGasPrice       *big.Int
-		originalLimit          uint64
+		originalLimit          uint32
 		limitMultiplierPercent float32
 		expectedLimit          uint64
 	}{
@@ -158,7 +158,7 @@ func Test_BumpDynamicFeeOnly(t *testing.T) {
 		bumpWei                *big.Int
 		maxGasPriceWei         *big.Int
 		expectedFee            gas.DynamicFee
-		originalLimit          uint64
+		originalLimit          uint32
 		limitMultiplierPercent float32
 		expectedLimit          uint64
 	}{
