+++ core/services/ocrcommon/transmitter_test.go
@@ -22,7 +22,7 @@ func Test_Transmitter_CreateEthTransaction(t *testing.T) {
 
 	_, fromAddress := cltest.MustInsertRandomKey(t, ethKeyStore, 0)
 
-	gasLimit := uint64(1000)
+	gasLimit := uint32(1000)
 	toAddress := testutils.NewAddress()
 	payload := []byte{1, 2, 3}
 	txm := new(txmmocks.TxManager)
