+++ core/services/pipeline/task.eth_tx.go
@@ -132,7 +132,7 @@ func (t *ETHTxTask) Run(_ context.Context, lggr logger.Logger, vars Vars, inputs
 		FromAddress:    fromAddr,
 		ToAddress:      common.Address(toAddr),
 		EncodedPayload: []byte(data),
-		GasLimit:       uint64(gasLimit),
+		GasLimit:       uint32(gasLimit),
 		Meta:           txMeta,
 		Strategy:       strategy,
 		Checker:        transmitChecker,
