+++ core/services/pipeline/task.eth_tx_test.go
@@ -68,7 +68,7 @@ func TestETHTxTask(t *testing.T) {
 				from := common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")
 				to := common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF")
 				data := []byte("foobar")
-				gasLimit := uint64(12345)
+				gasLimit := uint32(12345)
 				jid := int32(321)
 				addr := common.HexToAddress("0x2E396ecbc8223Ebc16EC45136228AE5EDB649943")
 				txMeta := &txmgr.EthTxMeta{
@@ -119,7 +119,7 @@ func TestETHTxTask(t *testing.T) {
 				from := common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")
 				to := common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF")
 				data := []byte("foobar")
-				gasLimit := uint64(12345)
+				gasLimit := uint32(12345)
 				txMeta := &txmgr.EthTxMeta{
 					JobID:         &jid,
 					RequestID:     &reqID,
@@ -153,7 +153,7 @@ func TestETHTxTask(t *testing.T) {
 				"fromAddrs": []common.Address{common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")},
 				"toAddr":    common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF"),
 				"data":      []byte("foobar"),
-				"gasLimit":  uint64(12345),
+				"gasLimit":  uint32(12345),
 				"requestData": map[string]interface{}{
 					"jobID":         int32(321),
 					"requestID":     common.HexToHash("0x5198616554d738d9485d1a7cf53b2f33e09c3bbc8fe9ac0020bd672cd2bc15d2"),
@@ -166,7 +166,7 @@ func TestETHTxTask(t *testing.T) {
 				from := common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")
 				to := common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF")
 				data := []byte("foobar")
-				gasLimit := uint64(12345)
+				gasLimit := uint32(12345)
 				txMeta := &txmgr.EthTxMeta{
 					JobID:         &jid,
 					RequestID:     &reqID,
@@ -200,7 +200,7 @@ func TestETHTxTask(t *testing.T) {
 				"fromAddrs": []common.Address{common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")},
 				"toAddr":    common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF"),
 				"data":      []byte("foobar"),
-				"gasLimit":  uint64(12345),
+				"gasLimit":  uint32(12345),
 				"requestData": map[string]interface{}{
 					"jobID":         int32(321),
 					"requestID":     common.HexToHash("0x5198616554d738d9485d1a7cf53b2f33e09c3bbc8fe9ac0020bd672cd2bc15d2"),
@@ -213,7 +213,7 @@ func TestETHTxTask(t *testing.T) {
 				from := common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")
 				to := common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF")
 				data := []byte("foobar")
-				gasLimit := uint64(12345)
+				gasLimit := uint32(12345)
 				txMeta := &txmgr.EthTxMeta{
 					JobID:         &jid,
 					RequestID:     &reqID,
@@ -250,7 +250,7 @@ func TestETHTxTask(t *testing.T) {
 				from := common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")
 				to := common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF")
 				data := []byte("foobar")
-				gasLimit := uint64(12345)
+				gasLimit := uint32(12345)
 				txMeta := &txmgr.EthTxMeta{FailOnRevert: null.BoolFrom(false)}
 				keyStore.On("GetRoundRobinAddress", testutils.FixtureChainID, from).Return(from, nil)
 				txManager.On("CreateEthTransaction", txmgr.NewTx{
@@ -293,7 +293,7 @@ func TestETHTxTask(t *testing.T) {
 					FromAddress:    from,
 					ToAddress:      to,
 					EncodedPayload: data,
-					GasLimit:       uint64(drJobTypeGasLimit),
+					GasLimit:       drJobTypeGasLimit,
 					Meta:           txMeta,
 					Strategy:       txmgr.SendEveryStrategy{},
 				}).Return(txmgr.EthTx{}, nil)
@@ -329,7 +329,7 @@ func TestETHTxTask(t *testing.T) {
 					FromAddress:    from,
 					ToAddress:      to,
 					EncodedPayload: data,
-					GasLimit:       uint64(specGasLimit),
+					GasLimit:       specGasLimit,
 					Meta:           txMeta,
 					Strategy:       txmgr.SendEveryStrategy{},
 				}).Return(txmgr.EthTx{}, nil)
@@ -351,7 +351,7 @@ func TestETHTxTask(t *testing.T) {
 				"fromAddrs": []common.Address{common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")},
 				"toAddr":    common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF"),
 				"data":      []byte("foobar"),
-				"gasLimit":  uint64(12345),
+				"gasLimit":  uint32(12345),
 				"requestData": map[string]interface{}{
 					"jobID":         int32(321),
 					"requestID":     common.HexToHash("0x5198616554d738d9485d1a7cf53b2f33e09c3bbc8fe9ac0020bd672cd2bc15d2"),
@@ -383,7 +383,7 @@ func TestETHTxTask(t *testing.T) {
 				from := common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c")
 				to := common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF")
 				data := []byte("foobar")
-				gasLimit := uint64(12345)
+				gasLimit := uint32(12345)
 				txMeta := &txmgr.EthTxMeta{
 					JobID:         &jid,
 					RequestID:     &reqID,
@@ -511,7 +511,7 @@ func TestETHTxTask(t *testing.T) {
 				"fromAddr":      common.HexToAddress("0x882969652440ccf14a5dbb9bd53eb21cb1e11e5c"),
 				"toAddr":        common.HexToAddress("0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF"),
 				"data":          []byte("foobar"),
-				"gasLimit":      uint64(12345),
+				"gasLimit":      uint32(12345),
 				"jobID":         int32(321),
 				"requestID":     common.HexToHash("0x5198616554d738d9485d1a7cf53b2f33e09c3bbc8fe9ac0020bd672cd2bc15d2"),
 				"requestTxHash": common.HexToHash("0xc524fafafcaec40652b1f84fca09c231185437d008d195fccf2f51e64b7062f8"),
